+++ R/substr2.R
@@ -1,4 +1,4 @@
-## Copyright (C) 2020  Brodie Gaslam
+## Copyright (C) 2021  Brodie Gaslam
 ##
 ## This file is part of "fansi - ANSI Control Sequence Aware String Functions"
 ##
@@ -68,7 +68,11 @@
 #'   interpreted, particularly if you are getting unexpected results.
 #' @param x a character vector or object that can be coerced to character.
 #' @param type character(1L) partial matching `c("chars", "width")`, although
-#'   `type="width"` only works correctly with R >= 3.2.2.
+#'   `type="width"` only works correctly with R >= 3.2.2.  With "width", whether
+#'   C0 and C1 are treated as zero width may depend on R version and locale in
+#'   addition what the `ctl` parameter is set to.  For example, for R4.1 in
+#'   UTF-8 locales C0 and C1 will be zero width even if the value of `ctl` is
+#'   such that they wouldn't be so in other circumstances.
 #' @param round character(1L) partial matching
 #'   `c("start", "stop", "both", "neither")`, controls how to resolve
 #'   ambiguities when a `start` or `stop` value in "width" `type` mode falls
@@ -96,11 +100,11 @@
 #'   assumptions `fansi` makes about how strings are rendered on your display
 #'   to be incorrect, for example by moving the cursor (see [fansi]).
 #' @param term.cap character a vector of the capabilities of the terminal, can
-#'   be any combination "bright" (SGR codes 90-97, 100-107), "256" (SGR codes
+#'   be any combination of "bright" (SGR codes 90-97, 100-107), "256" (SGR codes
 #'   starting with "38;5" or "48;5"), and "truecolor" (SGR codes starting with
-#'   "38;2" or "48;2"). Changing this parameter changes how `fansi` interprets
-#'   escape sequences, so you should ensure that it matches your terminal
-#'   capabilities. See [term_cap_test] for details.
+#'   "38;2" or "48;2"). Changing this parameter changes how `fansi`
+#'   interprets escape sequences, so you should ensure that it matches your
+#'   terminal capabilities. See [term_cap_test] for details.
 #' @examples
 #' substr_ctl("\033[42mhello\033[m world", 1, 9)
 #' substr_ctl("\033[42mhello\033[m world", 3, 9)