+++ src/unhandled.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020  Brodie Gaslam
+ * Copyright (C) 2021  Brodie Gaslam
  *
  * This file is part of "fansi - ANSI Control Sequence Aware String Functions"
  *
@@ -53,7 +53,7 @@ SEXP FANSI_unhandled_esc(SEXP x, SEXP term_cap) {
     SEXP chrsxp = STRING_ELT(x, i);
 
     if(chrsxp != NA_STRING && LENGTH(chrsxp)) {
-      FANSI_check_enc(chrsxp, i);
+      FANSI_check_chrsxp(chrsxp, i);
       const char * string, * string_start;
 
       string = string_start = CHAR(chrsxp);