+++ src/unique.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020  Brodie Gaslam
+ * Copyright (C) 2021  Brodie Gaslam
  *
  * This file is part of "fansi - ANSI Control Sequence Aware String Functions"
  *
@@ -41,6 +41,7 @@ SEXP FANSI_unique_chr(SEXP x) {
     for(R_xlen_t i = 1; i < x_len; ++i) {
       SEXP x_cur;
       x_cur = STRING_ELT(x_srt, i);
+      FANSI_check_chrsxp(x_cur, i);
       if(x_prev != x_cur) {
         ++u_count;
         x_prev = x_cur;