+++ tests/unitizer/utf8.R
@@ -238,9 +238,8 @@ unitizer_sect("Corner cases", {
   substr_ctl(b.t.c, 4, 4)
 
   substr2_ctl(b.t.c, 0, 0, type='width')
-  ## Commented out due to bug introduced by r79799 in r-devel
-  # substr2_ctl(b.t.c, 0, 2, type='width')
-  # substr2_ctl(b.t.c, 1, 4, type='width')
+  substr2_ctl(b.t.c, 0, 2, type='width')
+  substr2_ctl(b.t.c, 1, 4, type='width')
   substr2_ctl(b.t.c, 0, 5, type='width')
   substr2_ctl(b.t.c, 5, 5, type='width')
 })
@@ -287,9 +286,15 @@ unitizer_sect("nchar", {
 
   # _sgr
 
-  esc.4 <- c(sprintf("\033[31m%s\thello", w1), NA, hello.illegal)
+  esc.4 <- c(sprintf("\033[31m%shello", w1), NA, hello.illegal)
   nchar_sgr(esc.4, type='width', keepNA=FALSE, warn=FALSE, allowNA=TRUE)
   nzchar_sgr(esc.4, keepNA=FALSE, warn=FALSE)
+
+  # _sgr does not strip C0; note R behavior on width of C0-C1
+  # fluctuating around R4.1 transition so can't test directly.
+
+  nchar_sgr("\033[31m\thello", type='width') >=
+    nchar_ctl("\033[31m\thello", type='width')
 })
 unitizer_sect("unhandled", {
   # a bad utf8 string and other bad stuff
@@ -416,7 +421,7 @@ unitizer_sect("issue 54 ctd", {
     "#   time \033[3m\033[38;5;246m<list>\033[38;5;246m\033[23m, gc \033[3m\033[38;5;246m<list>\033[38;5;246m\033[23m\033[39m"
   )
   Encoding(string3) <- "UTF-8"
-  fansi::sgr_to_html(string3)
+  fansi::sgr_to_html(html_esc(string3))
 
   # head <- "<html><head><meta charset='utf-8'/></head><pre>"
   # f <- paste0(tempfile(), ".html")